[move-only] Borrow copyable typed arguments contained within a move only struct/enum.
Otherwise, we down the normal load [copy] path which will cause us to have a
tight exclusivity scope.

One thing to note is that if one accesses a field of a moveonly class one will
still get the tight exclusivity scope issue since SILGenLValue will emit the
moveonly class as a base of the field causing us to use the non-borrow codegen.

rdar://102746971