[Concurrency] TaskPool implementation
A TaskPool is similar to a TaskGroup, but functions very differently.
Tasks added to the pool execute the same way as in a group, however they
cannot be awaited on explicitly. The implementation does not track ready
tasks at all and therefore is not able to track the exact number of
times a next() would have to be resumed.

Instead, it automatically removes completed child tasks, and removes
their pending counts from the pool status. This allows us to implement a
waitAll() however it is not possible to wait for individual completions
(one could "wait for the next completion", but that is not very useful,
as we would not know _which_ one it was).

A TaskPool is useful in "runs forever" core loops of rpc and http
servers, where the top level task of the server runs a TaskPool, and
dispatches new tasks to handle each incoming request; and these may be
running concurrently. We want to use child tasks to keep this efficient,
but using a Group this is not possible because we must collect the
tasks as soon as they complete, but cannot do so from the task group
since it is a single task awaiting on a sequence.