[TypeLowering] Define "copy into" for opaque vals.
When building with opaque values enable, types which would otherwise get
AddressOnlyTypeLowering instead get OpaqueValueTypeLowering.  When such
types need to be copied into an address, the emitCopyInto method gets
called on the OpaqueValueTypeLowering.  So it must be implemented.

Additionally, vary its implementation based on whether the module is
address-lowered.  If it's not address-lowered, emit a copy-into as if
the type were loadable.  If it is address-lowered, emit a copy-into as
if the type were address-only.