build: disable bytecode layout library without stdlib
This directory is directly added from the top-level CMake which bypasses
the compiler swapping dance.  When cross-compiling or building the
toolchain with a non-clang compiler, the forced flags for the standard
library directory will result in build failures.  Avoid building the
library unless the standard library is being built (which will require
that the compiler-swap occurs).