Revert "Support cross compilation in macOS toolchain builds (#62306)"
This reverts commit a36ab6e287afc6c212ad250bfed440d5baf42b16.

The change to support cross compilation of `docc` in macOS toolchains
is breaking nightly toolchain builds – this will unblock them.

It looks like the `--test-product` option Swift-DocC's build script passes to
SwiftPM is supported when building for multiple architectures but not when
testing. PR toolchain builds don't run tests which is why this wasn't caught
before merging.

Resolves rdar://102908495