[sil] Refactor FieldSensitiveAddressPrunedLiveness along the lines of MultiDefPrunedLiveRange.
To do this I removed isWithinBoundary and friends from
FieldSensitiveAddressPrunedLiveness and moved them and defs into a new class
called FieldSensitiveAddressPrunedLiveRange. The former is now akin to
PrunedLiveness and the later is akin to MultiDefPrunedLiveRange. Since we are
dealing with addresses which are inherently not in SSA, I eliminated the extra
CRTP class that PrunedLiveness used to handle both SSA/MultiDef code.

I also ripped out FieldSensitiveAddressPrunedLivenessBoundary since as I was
experimenting with it, I realized that I fundamentally needed a different API
than what the non field sensitive case supports since I have to handle
TypeTreeSpans.