[ClosureLifetimeFixup] Shrink arg lifetimes.
Rather than inserting dealloc_stacks at the end of blocks in the
dominance frontier of the alloc_stack, walk backward in the block to the
last user (if any) and insert after that (or at the beginning of the
block if not).  If the transitive uses aren't understood, just insert at
the end of the blocks as before.