[OpaqueValues] Fix @in_guaranteed loadable yields.
When opaque values are enabled, when yielding a value with an indirect
convention, a value whose type's category is "object" must be yielded.
Previously, if a value's type was loadable, an attempt was made to yield
a value whose type's category was "address".  Here, that's fixed.
Additionally, handling for trivial types yielded via @in_guaranteed is
added.