[move-only-object] Use closureCapture to emit more informative diagnostic on lets being captured in a closure.
Specifically, previously we said something along the lines of...

'x' has guaranteed ownership but was consumed

even for lets. This is pretty misleading. Using @closureCapture, we can now emit
a much better diagnostic:

'x' consumed in closure. This is illegal since if the closure is invoked more than once the binding will be uninitialized on later invocations.

I am also going to use this technique to emit a similar diagnostic for vars that
are captured by a closure. It will be slightly different since vars will behave
like inout parameters.