[AddressLowering] Rematerialize projections eagerly.
The pass rewrites opaque values in reverse post order of their
definitions.  When an opaque value is rewritten, both its definition and
its uses are rewritten.  When a def or a use is rewritten, in order to
materialize an address with which to rewrite, projection instructions
are created.  Some of these projection instructions may be reused when
rewriting later opaque values.

As a result, it's possible to have two opaque values A and B (which are
rewritten in that order) such that rewriting a use of A which occurs
"after" the def (or a use) of B creates a projection P which is then
used by that "earlier" def (or use) of B:

    ```
    A =
    B =     // relies on P
    use B
    use A   // creates some projection P
    ```

When rewriting the def (or that use, respectively) of B, the projection
which was created for the use of A will be reused.  But that projection
instruction will be "after" the place where it is used when rewriting
the def or use of B.  That's invalid!

To address this, rather than reusing the old instruction, always create a
new instruction for a projection.  If the new instruction occurs
"before" (i.e. dominates) the old instruction replace uses of the old
with the new and delete the old.  Otherwise, just delete the new
instruction.