[move-only] Make sure that we convert copy_value/load [copy] move only to their after error variant if we error early.
In sum, if we hit an early error that says that we found a situation that we
don't know how to handle, we were not converting copy_value ->
explicit_copy_value and load [copy] -> load_borrow + copy_value.

Discovered this issue while attempting to enable sil-verify-all on the move only
tests. Once I enable that on the relevant tests, this will be tested by
sil-verify-all not triggering. I have a few other things to fix before I can do
that though.